+++ CHANGELOG.md
@@ -20,6 +20,10 @@ CHANGELOG
  * Add argument `$target` to `ContainerBuilder::registerAliasForArgument()`
  * Deprecate registering a service without a class when its id is a non-existing FQCN
  * Allow multiple `#[AsDecorator]` attributes
+ * Handle returning arrays and config-builders from config files
+ * Handle declaring services using PHP arrays that follow the same shape as corresponding yaml files
+ * Deprecate using `$this` or its internal scope from PHP config files; use the `$loader` variable instead
+ * Deprecate XML configuration format, use YAML or PHP instead
 
 7.3
 ---